fix: searches work for alphabet of size 256

+++ src/fmindex-collection/search/SearchNg14.h
@@ -114,7 +114,7 @@ struct Search {
                 search_next<OnMatchL, OnMatchR>(newCur, e, blockIter+1);
             }
 
-            for (uint8_t i{1}; i < symb; ++i) {
+            for (size_t i{1}; i < symb; ++i) {
                 auto newCur = cursors[i];
 
                 if constexpr (Deletion) {
@@ -123,7 +123,7 @@ struct Search {
                 search_next<OnSubstituteL, OnSubstituteR>(newCur, e+1, blockIter+1); // as substitution
             }
 
-            for (uint8_t i(symb+1); i < Sigma; ++i) {
+            for (size_t i(symb+1); i < Sigma; ++i) {
                 auto newCur = cursors[i];
 
                 if constexpr (Deletion) {
