+++ src/fmindex-collection/search/SearchNg15.h
@@ -88,7 +88,7 @@ struct Search {
                 search_next<'M'>(newCur, e, blockIter+1);
             }
 
-            for (uint8_t i{1}; i < symb; ++i) {
+            for (size_t i{1}; i < symb; ++i) {
                 auto newCur = cursors[i];
 
                 if constexpr (Deletion) {
@@ -97,7 +97,7 @@ struct Search {
                 search_next<'S'>(newCur, e+1, blockIter+1); // as substitution
             }
 
-            for (uint8_t i(symb+1); i < Sigma; ++i) {
+            for (size_t i(symb+1); i < Sigma; ++i) {
                 auto newCur = cursors[i];
 
                 if constexpr (Deletion) {
