+++ src/fmindex-collection/search/SearchNg21V6.h
@@ -120,7 +120,7 @@ struct Search {
                 if (abort) return;
             }
 
-            for (uint8_t i{1}; i < symb; ++i) {
+            for (size_t i{1}; i < symb; ++i) {
                 auto newCur = cursors[i];
 
                 if constexpr (Deletion) {
@@ -131,7 +131,7 @@ struct Search {
                 if (abort) return;
             }
 
-            for (uint8_t i(symb+1); i < Sigma; ++i) {
+            for (size_t i(symb+1); i < Sigma; ++i) {
                 auto newCur = cursors[i];
 
                 if constexpr (Deletion) {
