+++ src/fmindex-collection/search/SearchPseudo.h
@@ -95,7 +95,7 @@ struct Search {
 
         // search substitution
         if (l[pos] <= e+1 and e+1 <= u[pos]) {
-            for (uint8_t i{1}; i < Sigma; ++i) {
+            for (size_t i{1}; i < Sigma; ++i) {
                 if (i == rank) continue;
                 search_hm(cursors[i], e+1, pos+1);
             }
